refactor: Entity 접미사 제거 및 패키지 이동
