+++ app/src/main/java/nextstep/github/data/services/GithubService.kt
@@ -1,10 +1,10 @@
 package nextstep.github.data.services
 
-import nextstep.github.data.entities.RepositoryEntity
+import nextstep.github.model.Repository
 import retrofit2.http.GET
 import retrofit2.http.Path
 
 interface GithubService {
     @GET("orgs/{organization}/repos")
-    suspend fun getRepositories(@Path("organization") organization: String): List<RepositoryEntity>
+    suspend fun getRepositories(@Path("organization") organization: String): List<Repository>
 }
