+++ app/src/main/java/nextstep/github/model/Repository.kt
@@ -1,10 +1,10 @@
-package nextstep.github.data.entities
+package nextstep.github.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RepositoryEntity(
+data class Repository(
 
     @SerialName("full_name")
     val fullName: String?,
